use size_t instead og static cast

+++ sources/Runner/ScoreScene.cpp
@@ -58,7 +58,7 @@ namespace BBM
 		scene->addEntity("scene title text")
 			.addComponent<PositionComponent>(1920 / 2.37, 250, 0)
 			.addComponent<Drawable2DComponent, RAY2D::Text>("CONGRATS", 50, RAY::Vector2(), ORANGE);
-		for (int i = 0; i < static_cast<int>(players.size()); i++) {
+		for (size_t i = 0; i < players.size(); i++) {
 			scene->addEntity("player tile")
 				.addComponent<PositionComponent>(224 * (i + 1) + 200 * i, 1080 / 2.5, 0)
 				.addComponent<Drawable2DComponent, RAY2D::Rectangle>(RAY::Vector2(224 * (i + 1) + 200 * i, 1080 / 3),
